+++ tengine/game/entities/kinematic.cpp
@@ -8,14 +8,14 @@ NETVAR_TABLE_BEGIN(CKinematic);
 NETVAR_TABLE_END();
 
 SAVEDATA_TABLE_BEGIN_EDITOR(CKinematic);
-	SAVEDATA_DEFINE_HANDLE(CSaveData::DATA_COPYTYPE, float, m_flLerpTime, "LerpTime");
-	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, float, m_flLerpStart);
-	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, float, m_flLerpEnd);
+	SAVEDATA_DEFINE_HANDLE(CSaveData::DATA_COPYTYPE, double, m_flLerpTime, "LerpTime");
+	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, double, m_flLerpStart);
+	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, double, m_flLerpEnd);
 	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, Vector, m_vecLerpStart);
 	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, Vector, m_vecLerpGoal);
-	SAVEDATA_DEFINE_HANDLE(CSaveData::DATA_COPYTYPE, float, m_flAngleLerpTime, "AngleLerpTime");
-	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, float, m_flAngleLerpStart);
-	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, float, m_flAngleLerpEnd);
+	SAVEDATA_DEFINE_HANDLE(CSaveData::DATA_COPYTYPE, double, m_flAngleLerpTime, "AngleLerpTime");
+	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, double, m_flAngleLerpStart);
+	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, double, m_flAngleLerpEnd);
 	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, EAngle, m_angLerpStart);
 	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, EAngle, m_angLerpGoal);
 	SAVEDATA_DEFINE(CSaveData::DATA_COPYTYPE, bool, m_bLerping);
@@ -56,8 +56,8 @@ void CKinematic::Think()
 
 	if (m_flLerpStart > 0)
 	{
-		float flTime = GameServer()->GetGameTime() - m_flLerpStart;
-		float flLerp = RemapVal(flTime, 0, m_flLerpEnd - m_flLerpStart, 0, 1);
+		double flTime = GameServer()->GetGameTime() - m_flLerpStart;
+		float flLerp = RemapVal((float)flTime, 0, (float)(m_flLerpEnd - m_flLerpStart), 0, 1);
 		float flRamp = SLerp(flLerp, 0.2f);
 
 		m_bLerping = true;
@@ -70,8 +70,8 @@ void CKinematic::Think()
 
 	if (m_flAngleLerpStart > 0)
 	{
-		float flTime = GameServer()->GetGameTime() - m_flAngleLerpStart;
-		float flLerp = RemapVal(flTime, 0, m_flAngleLerpEnd - m_flAngleLerpStart, 0, 1);
+		double flTime = GameServer()->GetGameTime() - m_flAngleLerpStart;
+		float flLerp = RemapVal((float)flTime, 0, (float)(m_flAngleLerpEnd - m_flAngleLerpStart), 0, 1);
 		float flRamp = SLerp(flLerp, 0.2f);
 
 		m_bLerping = true;
@@ -108,7 +108,7 @@ void CKinematic::LerpTo(const eastl::vector<tstring>& sArgs)
 	m_flLerpEnd = m_flLerpStart + m_flLerpTime;
 
 	m_vecLerpStart = GetGlobalOrigin();
-	m_vecLerpGoal = Vector(stof(sArgs[0]), stof(sArgs[1]), stof(sArgs[2]));
+	m_vecLerpGoal = Vector((float)stof(sArgs[0]), (float)stof(sArgs[1]), (float)stof(sArgs[2]));
 }
 
 void CKinematic::LerpAnglesTo(const eastl::vector<tstring>& sArgs)
@@ -125,5 +125,5 @@ void CKinematic::LerpAnglesTo(const eastl::vector<tstring>& sArgs)
 	m_flAngleLerpEnd = m_flAngleLerpStart + m_flAngleLerpTime;
 
 	m_angLerpStart = GetGlobalAngles();
-	m_angLerpGoal = EAngle(stof(sArgs[0]), stof(sArgs[1]), stof(sArgs[2]));
+	m_angLerpGoal = EAngle((float)stof(sArgs[0]), (float)stof(sArgs[1]), (float)stof(sArgs[2]));
 }
